services/hdbpp-cluster-reporting/requirements.txt
pyyaml
apscheduler
flask
flask-restful
requests
Flask-RestPlus
Flask-SQLAlchemy
gunicorn